add note all command

+++ basement_bot/plugins/who.py
@@ -1,7 +1,9 @@
 import datetime
+import io
 
 import base
 import discord
+import yaml
 from discord.ext import commands
 
 
@@ -31,7 +33,7 @@ class Who(base.BaseCog):
     async def whois_user(self, ctx, user: discord.Member):
         embed = discord.Embed(
             title=f"User info for {user}",
-            description="**Note: this user is a bot!**" if user.bot else "",
+            description="**Note: this is a bot account!**" if user.bot else "",
         )
 
         embed.set_thumbnail(url=user.avatar_url)
@@ -43,11 +45,14 @@ class Who(base.BaseCog):
 
         user_notes = await self.get_notes(user, ctx.guild)
 
+        total_notes = 0
         if user_notes:
+            total_notes = len(user_notes)
             user_notes = user_notes[-3:]
+        embed.set_footer(text=f"{total_notes} total notes")
 
         for note in user_notes:
-            author = ctx.guild.get_member(int(note.author_id))
+            author = ctx.guild.get_member(int(note.author_id)) or "<Not found>"
             embed.add_field(
                 name=f"Note (from {author} at {note.updated})",
                 value=f"*{note.body}*" or "*None*",
@@ -95,10 +100,50 @@ class Who(base.BaseCog):
     async def clear_notes(self, ctx, user: discord.Member):
         notes = await self.get_notes(user, ctx.guild)
 
+        if not notes:
+            await self.bot.send_with_mention(ctx, "There are no notes for that user")
+            return
+
+        await self.bot.confirm(
+            ctx,
+            f"Are you sure you want to clear {len(notes)} notes?",
+            delete_after=True,
+        )
+
         for note in notes:
             await note.delete()
 
         await self.bot.send_with_mention(ctx, "I cleared all the notes for that user")
+
+    @note.command(
+        name="all",
+        brief="Gets all notes for a user",
+        description="Gets all notes for a user instead of just new ones",
+        usage="@user",
+    )
+    async def all_notes(self, ctx, user: discord.Member):
+        notes = await self.get_notes(user, ctx.guild)
+
+        if not notes:
+            await self.bot.send_with_mention(ctx, "There are no notes for that user")
+            return
+
+        note_output_data = []
+        for note in notes:
+            author = ctx.guild.get_member(int(note.author_id)) or "<Not found>"
+            data = {
+                "body": note.body,
+                "from": str(author),
+                "at": str(note.updated),
+            }
+            note_output_data.append(data)
+
+        yaml_file = discord.File(
+            io.StringIO(yaml.dump({"notes": note_output_data})),
+            filename=f"notes-for-{user.id}-{datetime.datetime.utcnow()}.yaml",
+        )
+
+        await self.bot.send_with_mention(ctx, file=yaml_file)
 
     async def get_notes(self, user, guild):
         user_notes = (
